functions/content/pyproject.toml
[tool.poetry]
name = "content"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
firebase-admin = "6.5.0"
google-cloud-storage = "2.19.0"
google-cloud-firestore = "2.21.0"
google-cloud-aiplatform = "1.113.0"
functions-framework = "3.9.2"
Pillow = "11.3.0"
requests = "2.32.5"
opencv-python = "4.12.0.88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
